- fix database layer to match requirement - make models to support insert and list items from database

+++ app/src/main/java/id/ac/ui/cs/mobileprogramming/nataprawiraf/kohi/data/model/Step.kt
@@ -9,5 +9,5 @@ data class Step(
     val stepId: Long,
     var name: String,
     var number: Int,
-    var timeInSeconds: Long
+    var timeInSeconds: Int
 )
